+++ app/src/main/java/org/hwyl/sexytopo/control/util/Space3DTransformerForElevation.java
@@ -16,7 +16,7 @@ public class Space3DTransformerForElevation  extends Space3DTransformer {
     }
 
 
-    protected void update(Space<Coord3D> space, Station station, Coord3D coord3D, double rotation) {
+    protected void update(Space<Coord3D> space, Station station, Coord3D coord3D, float rotation) {
         space.addStation(station, coord3D);
         for (Leg leg : station.getOnwardLegs()) {
             if (leg.hasDestination()) {
@@ -37,7 +37,7 @@ public class Space3DTransformerForElevation  extends Space3DTransformer {
             adjustedLeg = leg.adjustAzimuth(0);
         }
 
-        double delta = adjustedLeg.getAzimuth() - leg.getAzimuth();
+        float delta = adjustedLeg.getAzimuth() - leg.getAzimuth();
 
         Coord3D end = Space3DUtils.toCartesian(start, adjustedLeg);
         Line<Coord3D> line = new Line<>(start, end);
@@ -48,7 +48,7 @@ public class Space3DTransformerForElevation  extends Space3DTransformer {
     }
 
 
-    protected void updateSplay(Space<Coord3D> space, Leg leg, Coord3D start, double rotation) {
+    protected void updateSplay(Space<Coord3D> space, Leg leg, Coord3D start, float rotation) {
         Leg adjustedLeg = leg.rotate(rotation);
         Coord3D end = Space3DUtils.toCartesian(start, adjustedLeg);
         Line<Coord3D> line = new Line<>(start, end);
